Remove block_sizes from the Python data types.
We don't want to store the block_sizes array in the Python data type.  Why?
Several problems.  First, it's not really used in libparted that I can find,
so maybe it's a good candidate for removal.  Second, it is stored as an
array of ints, but with no NULL terminator and no count of how many elements
are in the array.  So how does that even work?  My guess is it doesn't because
it was never used elsewhere in the libparted code.