Merge the two Unknown*TypeExceptions into a single one.
We're probably going to have a lot of places where a parted function can get
an unknown disk type, filesystem type, or whatever.  It makes sense to just
have one exception for all these events instead of a whole pile.  The caller
will know what sort of thing is Unknown.