Moved methods to the object they correspond to
Reduced the _ped methods struct and moved most of the methods out to the corresponding object.  Removed the Py_InitModule3() calls in the init function for _ped because those were driving over our method struct.

Updated the to do list with some new ideas.  We have a problem with some of the natmath functions.  Not sure where to put those.  Also, we lack _ped_Unit now, but should we bring it back as a method container and nothing more?