Started PedDevice old-API class
Moved PedDevice class to the PedDevice module.  The old pyparted named the modules and classes the same, which makes it a bit confusing.  This is the start of the implementation of the deprecated API for PedDevice.  Visibility is key for these old modules, which will require some fun meta class tricks.