Use 'z' in PyArg_ParseTuple and fix ped_device_*_all() functions.
When reading strings with PyArg_ParseTuple(), use 'z' for the
format string so we get NULL if the string value is None in Python.

Fixed py_ped_device_probe_all() and py_ped_device_free_all() so
they do not expect any input parameters and always return None.
The functions in libparted take no parameters and are void.