Move higher level Device methods to _ped module.
With the Device type, we have determined that in order to match the
functionality of libparted, we need to prevent the Device type from
being instantiated, but allow the user to get a Device type when
calling ped_device_get().

The following functions have moved to the _ped module:
    _ped.Device.get() -> _ped.getDevice()
    _ped.Device.probe_all() -> _ped.probeAllDevices()
    _ped.Device.free_all() -> _ped.freeAllDevices()

Also fixed a double free that was occuring in py_ped_device_get().