Use Py_XDECREF() on PyMemberDef members that are PyObjects.
For every type we have, there exists a PyMemberDef structure that
defines the type members, which are attributes.  Everything is
ultimately a PyObject in the C API, which means that a PyMemberDef
struct can contain C primitives or other PyObjects.  Structure
members that are PyObjects need to have a Py_XDECREF() call made
on them in the type's dealloc function so that the Python machine
will clean up properly.