Bring _ped.Device closer to the way we want it to function.
That is, the type needs to exist and we need to be able to create
internally in _ped, but users cannot instantiate a new _ped.Device
and build it up programatically.  Here's where we are as of this
commit:

In [1]: import _ped

In [2]: a = _ped.get_device("/dev/sda")

In [3]: a.length
---------------------------------------------------------------------------
AttributeError                            Traceback (most recent call last)

/home/dcantrel/pyparted/<ipython console> in <module>()

AttributeError: '_ped.Device' object has no attribute 'length'

In [4]: a
Out[4]: <_ped.Device object at 0xb7b7bd40>

In [5]: a.length
Out[5]: 20480544L

In [6]: b = _ped.Device()
./localtest: line 3: 13230 Segmentation fault      ipython
[dcantrel@electron pyparted]$

You can see we have quite a bit more work to do.