Fix _ped.Device object creation so you can only create them from functions.
The tp_alloc function we had was causing segfaults, so I replaced it with
PyType_GenericAlloc which is what the python docs recommend we do anyway.
Also we were setting a tp_new in _pedmodule.c which was overriding our NULL
setting in the type object.  That'd explain why a = _ped.Device() still
worked.