In *_destroy() functions, destroy the Python object.
In our py_ped_*_destroy() pass-through functions, we were converting
the Python object to a libparted type, then destroying and returning
None.  While this works, it seems pointless.  We don't need to create
and destroy a libparted type, we only need to destroy our Python type,
so call PyObject_DEL() and return None.

The py_ped_file_system_clobber() function is special.  We have to
call the file system clobber function to actually destroy it on disk
and then destroy the object.