Fix PedAlignment2_ped_Alignment() missing ob_type problem.
This fixes the following problem:
    import _ped
    x = _ped.Alignment(10, 10)
    isinstance(x, _ped.Alignment) -> True
    y = x.duplicate()
    isinstance(y, _ped.Alignment) -> False

The problem was because we have our PyTypeObject declarations in the
header files to the project and have them all set to 'static', which
means we have too many copies.  Only the one seen in _pedmodule.c was
getting ob_type set correctly, which is why x is an instance of
_ped.Alignment, but not y.

The solution is to remove the static part of the declaration and put
it in the appropriate source file in the project, but then add this
line to the header file:

    extern PyTypeObject _ped_Alignment_Type_obj;

The extern line is so we can use that type object throughout the rest
of the code.