Created include/typeobjects/*.h to hold PyTypeObject definitions.
I like header files for source organization.  Since I did that for
the docstrings, I decided to move the PyTypeObject stuff in to
header files under the include/typeobjects/ subdirectory.  That way,
we can work on those structs in one location and not end up with
huge source files in src/