Fix converting libparted execptions into python exceptions.
First, we were always hitting a double free because libparted frees the error
string for us, but we were also doing that in partedExnHandler.  Second, the
exception flag and message need to be moved into a .c file and externed in
the header, just like was required for fixing the isinstance problem.