Copy in missing members in _ped_Partition_init()
For some objects, we only require the user to give us a bare minimum
of information and then we let libparted compute the other members.
To get these additional members back in to the Python object, I was
doing a Py_XDECREF on self and then simply converting the newly
generated libparted type to a Python object and assigning it to self.
On paper, this sounded like a good idea.  Several problems:

    1) Py_XDECREF only decrements the count by one.  We really need
       to zero out the count.
    2) Clearing self and then assigning something new to it while
       you're inside the object isn't very clean.
    3) The new self needed a Py_INCREF.

Things I've learned:

    - PyArg_ParseTuple and PyArg_ParseTupleAndKeywords do not run
      Py_INCREF for you on PyObjects it reads.

The new solution is pretty simple to follow.  We keep the parsing
using PyArg_ParseTupleAndKeywords and we keep the libparted call
because that computes the other members we need, but changes I've
made:

    1) Py_INCREF the PyObjects read by PyArg_ParseTupleAndKeywords
    2) Copy in computed members that are not PyObjects by just
       using an assignment in C.
    3) For computed members that are PyObjects, first I Py_CLEAR
       the member and set it to NULL, then I convert the computed
       type to a PyObject, and lastly I call Py_INCREF on that
       member.
    4) Use ped_free() to destroy the PedPartition we made.

In short, we need to use Py_INCREF on all PyObjects read by the
PyArg_ParseTuple functions.  We should not clear self and reconvert
the object, instead we should correctly copy in the new members.