Use tp_alloc for types that must use factory functions.
We have some types that users are not allowed to create on their
own.  The module may hand them one of these object types, but one
cannot be created in the usual way.

For these types, we have tp_new set to NULL.  In convert.c, make
sure we call tp_alloc rather than tp_new to allocate the new object
and return it to the caller.