Initialize members in _ped_Geometry_init() correctly.
Call Py_INCREF() on the _ped.Device object.  For start, length, and
end, just copy those back from the PedGeometry we get.  end may or
may not be provided by the user, but it doesn't matter because we
should be able to just copy the values directly back from the
PedGeometry and know we have everything we need.  They are long long
values in our struct and in libparted, so no special PyObject
handling for these.