Handle object members in _ped.FileSystem correctly.
This object may need some more work later.  We allow users to create
_ped.FileSystem objects, but we don't want to call libparted from
inside the __init__ function to see that what the user provided is
valid.  Why not?  The only function we get in libparted to check
that is the ped_file_system_create() function, but the user may not
want to create the filesystem on disk yet.

The disk and type parameters are required, so make sure to Py_INCREF
those correctly.  An optional parameter for now is checked.  Init it
to false.  I say this function may need some more work because we
may want to change 'checked' to be a boolean.