Fix errors in converting _ped.Device to and from PedDevice.
First, _ped_Device2PedDevice was only doing a ped_device_get, it wasn't
actually copying any of the values into the PedDevice object.  That's not
going to work.  Second, we have to copy the arch_specific pointer around
since that has function pointers that libparted methods will need to
access.