Store PedDevice pointer in _ped_Device struct.
The PedDevice structs in libparted are unique.  Unlike the other
types in libparted, PedDevices are generated when the library
loads and scans for everything.  Things like Constraint and
Alignment are can created on the fly and thrown away.  But Devices
are arch-specific and need to live through the life of the program.

In our _ped_Device struct, keep our convenience copies of the data,
but also store a pointer to the PedDevice we're mirroring to
Python so that we can use that in other calls where we need the
PedDevice.

This solves the problem of "what to do about the arch_specific data"
issue.