Do not run ped_device_destroy() all over the place.
Do not use ped_device_destroy() to free a PedDevice.  We are not
working with PedDevice types like we do with a PedGeometry or
something similar.  We are passing the pointers around to the
master copies of PedDevice objects.  If you are getting a
PedDevice by calling _ped_Device2PedDevice(), you are getting a
pointer and you SHOULD NOT destroy it.

The only place where we should call ped_device_destroy() is in
py_ped_device_destroy().