Update parted.PedDevice for changes to _ped.Device.
It's been a long time since we looked at the parted module.  One big thing
we changed is that you can no longer create a _ped.Device yourself.  We also
moved a couple methods around.  Update PedDevice to reflect these changes.
I have not yet tested this module, but test cases for it will be coming.