Move DiskType get functions to the top level _ped module.
DiskTypes are initialized at runtime, but are essentially static.
We can remove a type once the library is loaded, but we never really
create new types.  They are merely for reference purposes to give us
things like type names and feature lists of disk types.

I moved the get functions for DiskType to _ped so that in the
parted/disk.py module, I can initialize a diskType hash table that
contains all of the disk types with the type name as the key value
of the hash.