Add parted.FileSystem class.
This one will need some more work, but the framework is there.

NOTE:  The probe and probe_specific functions in libparted exist
at the parted level rather than being methods on a FileSystem.
Additionally, the get_create_constraint and get_copy_constraint
functions in libparted are more closely related to a Device than
a FileSystem, so they exist as methods on a Device.