Put get_create_constraint() and get_copy_constraint() on _ped.Device.
Put the get_create_constraint() and get_copy_constraint() as
methods on the _ped.Device object.  We originally had these as methods
on _ped.FileSystem because they are functions in filesys.c in
libparted.  However, looking at the code in libparted, you can tell
it's more obvious that the action is on a specific device, not a
specific filesystem.

The code for these methods still lives in pyfilesys.c to keep it
consistent with the libparted code, but the methods in Python live
on the _ped.Device class.