Support keywords or no keywords in init, read PedSector as long long.
In _ped_Geometry_init(), rewrite the function to support cases where
there are no keywords provided.  For those instances, we want to
use PyArg_ParseTuple().  If there are keywords, we want to call
PyArg_ParseTupleAndKeywords().

For all instances where we are reading in a long long int using a
PyArg parsing function, use "L" for the format string.  We were
using "l", which leads to problems on 32-bit architectures.  All
PedSector values are long long ints and need the "L" conversion.