Fix PyArg argument processing for 32-bit architectures.
This was a fun one to track down.  After figuring out the code that
was failing on my i386 devel box (but works on clumens' x86_64 box),
I looked further and saw that our PedSector values were getting lost
by the time we hit the next function.

Anything that gets converted down to a PedSector needs to be treated
as a long long int.  We were converting those using "l" in the PyArg
parsing functions, which is just long int.  We need to use "L" to
make sure it always uses long long int.

For enum values, I changed those from long int to int.  enum values
are strange and we can probably leave them as long ints, but I want
to be specific and consistent across the code base.