Support having a filesystem type of None on Partition objects.
If a _ped.Partition object has a type of PARTITION_FREESPACE or
PARTITION_METADATA, it must not have an associated filesystem type.
Otherwise, you'll get strange assertions when checking if a partition is
active.  Now I just need to modify everywhere else we support optional
arguments to make sure they get set to a default.