Started the fdisk.py command
Started the fdisk.py command, which aims to be a more or less
compatible replacement for /sbin/fdisk that uses the pyparted
module and is written in Python.  We may not be able to do
everything that fdisk can do, but the point of this program is
to exercise the pyparted module.