Add getSize() method to parted.Partition
The getSize() method will return the size of the partition in a
unit suitable for human consumption.  Internally, parted deals
with sectors for the partition length, so we have to convert that
to something suitable for display.  partedUtils.py had a
getPartSizeMB() function to convert the length to megabytes.  This
new getSize() method defaults to megabytes, but can also handle
bytes through terabytes.  Adding new size conversions should be
easy too.