Import all objects, rename getDisk() to getDevice()
Fill out __all__ and make sure we import everything in to the same
namespace.

The getDisk() method was a bit of a misnomer.  It gives you a
parted.Device, so rename it to getDevice().

I've added some getSize() methods to Partition and Device objects,
which work sort of the same way.  Put the exponents hash at the top
level to share it between those methods.