Fix __init__ to accept right parameters, add getSize()
The __init__() methods takes either a path to a device on the
filesystem or an existing _ped.Device object.  Removed the device
parameter, since it didn't make sense.

Added the getSize() method to convert the size of the Device,
which is internally stored as a number of sectors, to a value
suitable for display to the user.  The conversion function
defaults to megabytes, but you can pick your unit from bytes
up to terabytes.