Make the device property on parted.Geometry give parted.Device
The Geometry class had a getPedDevice() method to return the
lowlevel _ped.Device object associated with this Geometry.
However, at the higher level code, we want to deal with
parted.Device as much as possible.

Make the 'device' attribute on parted.Geometry return a
parted.Device version of the internal _ped.Device.