Fix the geometry read and write test cases.
These were only failing because I didn't understand how the methods were
supposed to work.  If you try to write more bytes than the buffer contains,
it'll only write the contents of the buffer once.  It doesn't copy the
buffer over and over again into the geometry to meet the byte request.