Fix up _ped_Disk_init() and PedDisk2_ped_Disk()
Back when this function was first written, the thought was that
the user could optionally provide a disk type and we could then
call ped_disk_new_fresh() underneath to relabel the disk.  The
problem is this causes our conversion functions to fail and gets
us in to difficult situations in the code.

A _ped.Disk() now can only be initialized with a _ped.Device as
the argument.  ped_disk_new() is called.