Do not store PedDevice pointers in Python data structures.
We were doing this in order to prevent segfaults.  I think this
is causing more problems now.  Looking through libparted, we can
call ped_device_get() as many times as we want and it will give
us a pointer to the PedDevice in the device list that corresponds
to the path we give it.  I've modified the code to work this way
and got rid of the 'ped_device' struct members for _ped_Device
and _ped_Geometry.