Changed Disk._partitions to a list.
Was storing this as a hash table with the key being the partition
number and the value being the parted.Partition.  I thought that
would be a useful data structure, but in every instance that I've
been using Disk._partitions, I've only been using the values()
part, so change it to just a list of parted.Partition objects.