Keep the part_list and flags from PedDisk.
We are losing information when we convert to and from a _ped_Disk
and PedDisk.  We're using ped_disk_new(), which works fine to give
us the right struct, but we lose anything in memory that has not
been committed to the disk.

Save part_list, needs_clobber, and update_mode.  The block_sizes
struct member of PedDisk is not used at all in libparted, so we
just ignore it.