After ped_disk_add_partition() call, update partition number.
The ped_disk_add_partition() function in libparted takes in a
new PedPartition and adds it to the in-memory partition table
for the disk.  The partition is given a number during this
operation, so after the call we need to update the number in
our Python object otherwise every partition is -1 that we add.