Increase minor version number for next release.
Want to get in the habit of keeping the master branch at the next
release version number so I don't forget to bump at the end.  I am
not storing the spec file for the pyparted package in the upstream
project, so configure.ac is the location of the version number.