Do not keep a borrowed refence without INCREF-ing it (hdegoede).
In _ped_Disk_init we store a borrowed reference in self->dev (through
PyArg_ParseTuple()). We INCREF this on success at the end. But we keep it
without INCREF-ing it in an error path. This path clears the borrowed reference
in the error path. This fixes various cases of python crashing when having
an invalid partition table.