Add __str__ method for _ped.Constraint()
This one is more complex than the one for _ped.Alignment.  Since
a _ped.Constraint has two Alignments and two Geometries and two
Sectors, we need to call the repr() methods a few times.  Namely,
we need to call repr() for the two Alignments and the two
Geometries.  Sectors are long longs, so we can pull those in
directly.

Also stubbed out the test case.