Call commitToDevice() then commitToOS() in commit()
This is exactly what ped_disk_commit() does in libparted, but for
whatever reason, calling that function does not work reliably.  Tracing
the calls, the lines hit are identical.  The only thing that really
happens when we commit to OS is that libparted calls sync() a lot and
issues BLKRRPART.

In an effort to close out storage bugs, I'm changing this in pyparted.

https://bugzilla.redhat.com/show_bug.cgi?id=491746