Copy back values from ped_alignment_init()
Found this problem while working on test cases.  It's not sufficient for
us to just try to create a new alignment, we need to copy back the
possibly-adjusted offset and grain_size from the newly created
alignment.  ped_alignment_new() calls ped_alignment_init(), which in
turn calls abs_mod() if grain_size is non-zero and the offset of the
alignment is adjusted accordingly.  We need these adjusted values in our
_ped.Alignments.