Free string returned from libparted unit format functions.
The PedUnit functions that return strings depend on the caller to call
ped_free() to release the memory.  We take the string from libparted and
turn it in to a Python string, which makes a copy and lets Python memory
management take over.  Be sure to call ped_free() on the pointer
returned from libparted.