Read in PedUnit values from Python as ints consistently.
The pyunit.c code was using long and int inconsistently for reading in a
PedUnit value from Python, which would result in losing information
since the size of int and long differ.

Since PedUnit is just an enum in libparted, use int in _ped's code to
read and store PedUnit values.