Handle parted exceptions arising from ped_device_get (#495433).
The bug has an extensive description of what's going wrong, but the short
story is that by not handling the libparted exception, partedExnRaised stays
set so the next time a libparted function returns NULL, we think that
later caller is where the problem happened.  Not only does this raise an
exception where one shouldn't happen, but it raises the exception with a
very outdated and misleading error message.

We likely need to audit pyparted for other instances of this problem.