Use libparted commit() for parted.disk.Disk.commit() (hdegoede).
Currently we are using python code to combine the commit_to_dev and
commit_to_os to get commit() however libparted has a native function for this.
Using this removes an open / close of the device node and thus an udev change
event and unnecessary udev rule scanning. Note that for the unnecessary open /
close to be really gone we need a libparted with commit
ad25892bb995f61b0ddf801ed1f74e0b1e7390ce.

Note that this effectively backs out a previous patch to the commit function
that was needed because of errors in libparted.  With the above mentioned
libparted commit, this is fixed and we no longer need our workaround.

https://fedorahosted.org/pyparted/ticket/24