Use sectorSize not physicalSectorSize for size calculations
physicalSectorSize is the sector size on the platters of the disk,
where as sectorSize (which is the logical sector size) is the sector
size seen in the (s)ata / scsi commands and on higher levels.

IOW physical sector sizer is for informational purposes only, geometry,
alignments, etc. Are always expressed in logical sector sizes, so
we should use sectorSize in size calculations.