Put a deprecation warning in py_ped_constraint_duplicate().
Even though the goal of the _ped module is to create a complete Python
API of libparted, some public functions in libparted just seem
unnecessary in a Python module.  The duplicate() functions are among
these.  In Python we can do a copy.deepcopy() and get the same effect
that the libparted duplicate functions provide.  Put a deprecation
warning in this function to alert users that it will eventually go away
and they could just use a copy.deepcopy() instead.