Correct py_ped_file_system_probe_specific() for NULL returns.
If the filesystem you try to probe for is not found in the geometry
region specific, libparted returns NULL.  It may also raise an exception
from the filesystem code, but it does not seem appropriate to raise that
here.  I could be wrong, but I changed this function so on detection
failures, we just return None.  That seems more appropriate to the use
of this function in Python.