Add _pedmodule and parted functions for per disk flags
parted-2.1 and Fedora's parted-1.9.0-24, add functions for having per
disk flags (instead of per partition). These are used to enable / disable
cylinder alignment on disklabels where it is optional (the Fedora specific
patch which add the toggleAlignToCylinders function was not accepted
upstream).

This patch adds _pedmodule and parted module support for these flags,
this support is modelled after the partition flags and works identically.