Return PED_EXCEPTION_NO for yes/no interactive exceptions.
If we get a PED_EXCEPTION_YES_NO exception, we return PED_EXCEPTION_NO
to libparted to avoid any potential data loss ('yes' is assuming the
yes/no question would use the affirmative response to proceed with
destructive operations, which may always be the case...it is for now).
In pyparted, we then pass up an appropriate exception to the caller so
the user can be prompted and action taken.