Remove #! from test files, add $(PYTHON) to TESTS_ENVIRONMENT.
We had #!/usr/bin/python at the top of each test file that ran, which is
fine unless you are on a system with multiple versions of Python.
Removing the sha-bang line and add $(PYTHON) to the TESTS_ENVIRONMENT
variable in the Makefile.am files, which ensures our TESTS are run with
the Python picked up by the configure script.