Added getLength method.
The classes Device, Partition, and Geometry now have a getLength method that
is to replace the getSize method. It uses the new formatBytes function,
allowing for a wider (and more correct) range of byte prefixes. It defaults to
returning the length in sectors, so that it mirrors the length property. Also,
sectors can be passed as an argument to allow easier dynamic unit selection in
user programs.

Signed-off-by: David Cantrell <dcantrell@redhat.com>