Add support for PED_PARTITION_LEGACY_BOOT flag.
This is currently in the master branch for upstream parted and is part
of the parted package in Fedora rawhide.  The latest stable release of
parted is 2.3.  Since we cannot depend on the library version to know if
PED_PARTITION_LEGACY_BOOT is defined, add in a special check for that
constant and enable it in _ped and parted if it's present.