Correct boilerplate modifications.
The boilerplate in source files is not freeform.  Please adhere to
the format listed.  If the authors listing says "Red Hat Author(s)",
it is ok to change that to just "Author(s)" and add your name to the
listing.  Please adjust the copyright date range as well when you
make a signification modification to a source file.