Add support for hfs_esp flag
From: "Brian C. Lane" <bcl@redhat.com>

Fedoda's parted now has support for creating a HFS+ ESP for use on Mac
systems. This adds support for that flag which uses a new GUID for the
partition on GPT disklabels.

(This is an adjustment to bcl's original patch, which did not apply
cleanly against this branch.  The patch itself is relatively simple and
easy to follow.  The other three related to this did apply.)