Add function for resetting partition's number.
The partition's number is a read-only property because it doesn't make
sense to change it directly. However, it is sometimes necessary to reset
partition's number to default which tell's parted that the partition
should have a new number assigned. An example of such case is removing
partition from a disk and adding it back.

Related: rhbz#1166598
Signed-off-by: Vratislav Podzimek <vpodzime@redhat.com>