Add new functions to extend exception handling.
We've been avoiding this for an awful long time, but there is now a need for
users of pyparted to be able to specify what answers should be given to parted
exceptions that ask a question.  Specifically, we need to be able to ask the
user whether to proceed with a disk containing a corrupted GPT disk label or not.

This adds a function to register a callback that can prompt the user (among other
possible actions) and a function to clear this callback to restore default
behavior.