Beware of Python 3's version of the map() built-in function
It not only returns an iterator in Python 3, it actually works in the lazy
evaluation fashion. So if its returned value is not iterated over, the passed
function is not called on the iterable's items. Let's use a for cycle in these
places to make the iteration and function calls explicit.