Conditionally include new libparted flags
Previously #ifdef was used to check for new flags, but they are enum
symbols, not macros so all the ifdef's were failing and no new flags
were included.

This checks the value of PED_PARTITION_LAST_FLAG to see if newer flags
are defined. The only catch is that it must match the order in
parted/disk.h