Fix conditional checks for newer libparted symbols.
The preprocessor cannot evaluate the PED_PARTITION_LAST_FLAG when it is
defined as a enum symbol, and it certainly cannot compare it to enum
symbols that don't exist. So this unconditionally turns on the currently
supported libparted flags (msftdata, irst, esp).

This removes the nonfs flag support since upstream has declined to merge
support for it at this time.