Fix two problems found from running "make ci".
These are:

************* Module tests.test__ped_geometry
W1505(deprecated-method):377,8: GeometryStrTestCase.runTest: Using deprecated method assertRegexpMatches()
************* Module parted.cachedlist
W0221(arguments-differ): 85,4: CachedList.index: Arguments number differs from overridden 'index' method

The first is a python3 vs. python2 thing, so protect that with a check
for the version and tell pylint to ignore it when not appropriate.  The
second is because Sequence.index now takes extra arguments.  Accept
those in a way that shouldn't break on older versions.