Remove an old legacy function from anaconda's partedUtils.py (#41)
This function is invalid and should not be used.  It does not reflect
the current actual limits imposed by the kernel across different
platforms and disk labels.  Going forward, users should simply try to
add another logical partition and if it fails, handle the error
appropriately.  There is no way exposed in libparted to provide this
information, so pyparted will stop lying.