Make the partition name a property on parted.Partition objects (#34)
Some disk labels, such as GPT, support a partition name.  For labels
that do not support names, you get a PartitionException.  For supported
labels, you can get and set the name like this:

    print(self.part.name)
    self.part.name = "rootpart"

I have not exhaustively tested this, so if you use it, I would be
interested to receive feed back and/or bug reports.