Fix writing NULL sectors using a Geometry object() (#54)
This was a PyArg_ParseTuple() misuse.  Change it over to using s# to
read the buffer as an array of bytes, capture the size, then call
ped_geometry_write() in libparted.

Add a test case for this change too using the reporter's description
here:
    https://github.com/dcantrell/pyparted/issues/54

There may be more PyArg_ParseTuple() problems like this in the code.