pedmodule.c: Fix partition enum flag handling
The PED_PARTITION_* constants are enums, not defines. This means they
are not available for the preprocessor to evaluate with #if or #ifdef,
this includes the PED_PARTITION_FIRST_FLAG and _LAST_FLAG.

So in order to make pyparted somewhat flexible about which parted
version is is compiled against we need to evaluate the _LAST_FLAG at
runtime when adding the constants.