Confirm os.system commands run
os.system('mke2fs') assumes that it's installed (good bet), but also
that it works. On Ubuntu it doesn't, not with the FS sizes currently
encoded, and that leads to mystery test failures.

The straightforward way to fix this is just to raise the error where it
happens rather than silently fail; this uses asserts.