Bind low level APIs for type ID and UUID accessors
This binds the new methods:

  ped_partition_get_type_id
  ped_partition_set_type_id
  ped_partition_get_type_uuid
  ped_partition_set_type_uuid

There is no obvious compile time version info in the parted
headers that can be used to conditionalize the API usage.
Thus the code uses the PED_DISK_TYPE constants as a witness
for the APIs existing, on the basis that both were introduced
to the parted API concurrently.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>