Replace outdated Red Hat sample GPLv2+ boilerplates
When I started pyparted, Red Hat provided a set of templates to use
for license boilerplate text in source files.  I took this and used it
in this project, which is GPLv2+.  To me I thought it was just the
standard GPL boilerplate with the addition of "Red Hat Author(s)" to
indicate that Red Hat contributed to this project.  But upon closer
inspection, I see Red Hat added a clause about company trademarks
which has nothing to do with the software license.

At this point in time, Red Hat does not provide any such template and
discourages the use of long standing common GPL boilerplate text.
Instead preferring very simple comments like this:

    Copyright The PROJECT Authors
    SPDX-License-Identifier: SPDX-ID

You should also include an AUTHORS file for reference, but it's also
ok to not do that since the information is in the git project history.
It is not necessary to list each contributor as an author in the files
themselves.  Note the actual copyright statement is simply "Copyright
NAME".  There is no year, years, or year range specified.  There's
also no "(c)" or "(C)".  These are not necessary and legend has it
they showed up to make copyright statements look "more official".
Copyright statements just need to use the word Copyright, Copr., or
the © symbol.  I chose the word Copyright because it is very easily
displayable in a wide range of editors.  It is also worth noting here
that "(c)" and "(C)" are not valid replacements for the © symbol.

For more information:
    https://matija.suklje.name/how-and-why-to-properly-write-copyright-statements-in-your-code
    https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects

Signed-off-by: David Cantrell <dcantrell@redhat.com>