Help the _ped module garbage collection in the test suite
This showed up starting with Python 3.12, but the test suite was
segfaulting inconsistently and it looked to be tied to Python's
garbage collection routines.  The _ped module is written with garbage
collection support so in order to more explicitly assist Python with
the reference counting, set objects to None when we are done with them
in the test suite base classes.

Confirmed this works for both Python 3.11 and 3.12, so hopefully this
is good now.  This took an inordinate amount of time to track down and
I do not wish to do it again.  Famous last words, I guess.

Signed-off-by: David Cantrell <dcantrell@redhat.com>